refactor oneinstance
